requirements.txt
flask-cors
gunicorn
pip-chill
requests
